+++ web/documentserver-example/php/src/views/DocEditorView.php
@@ -231,23 +231,9 @@ final class DocEditorView extends View
             // encode the dataMailMergeRecipients object into the token
             $dataMailMergeRecipients["token"] = $jwtManager->jwtEncode($dataMailMergeRecipients);
         }
-        $out = getHistory($filename, $filetype, $docKey, $fileuri, $isEnableDirectUrl);
-        $history = $out[0];
-        $historyData = $out[1];
+        
         $historyLayout = "";
         if ($user->id != "uid-0") {
-            if ($history != null && $historyData != null) {
-                $historyLayout .= " config.events['onRequestHistory'] = function () {
-                        // show the document version history
-                        docEditor.refreshHistory(".json_encode($history).");};";
-                $historyLayout .= " config.events['onRequestHistoryData'] = function (event) {
-                        var ver = event.data;
-                        var histData = ".json_encode($historyData).";".
-                    "docEditor.setHistoryData(histData[ver - 1]);};
-                     config.events['onRequestHistoryClose'] = function () {
-                        document.location.reload();
-                    };";
-            }
             $historyLayout .= "// add mentions for not anonymous users
                 config.events['onRequestUsers'] = function () {
                     docEditor.setUsers({  // set a list of users to mention in the comments